requirements.txt
Django==1.8.3
djangorestframework==3.2.4
django-cors-headers==1.1.0
